Remove style handling from the terminal for now
Styling in the terminal requires a state machine of previously set
styles in order to correctly handle style resets. I haven't done that
yet, so to make the tests run again, I've removed it for now.

Another interesting challenge is going to be the HTML renderer, which I
can either handcode (probably the better solution), or build based on
lucid. The latter would require a lot of lookahead down the rendering
pipeline in order to find out what to actually put in e.g. a <strong>
pair, because in my rendering pipeline there is no explicit "this is
bold" instruction, only "the following is bold", hopefully eventually
followed by "okay revert the last style instruction".