Skip release checks on the nightly build
This should fix the issue that the checker scripts really only work with
the supplied stack.yaml file, and break if the Stack snapshot uses a
different compiler version, since the parameters aren’t passed properly.
Not a nice solution, but it works reasonably well. If this becomes a
problem, then the scripts should be rewritten in a better language than
fucking Bash.